Bug fix in SUM sample rate with float types. Added synchronization to the registers and made the data logger reset the samplings when it logs.

+++ src/ca/brood/softlogger/modbus/register/RealRegister.java
@@ -68,7 +68,7 @@ public class RealRegister extends Register implements Comparable<RealRegister>{
 			//Only for registers
 			samplingValue += temp.getFloat();
 			samplingCount ++;
-			this.setData((int)samplingValue);
+			this.setData((float)samplingValue);
 			break;
 		case MEAN:
 			samplingCount ++;
